requirements.txt
aiofiles==0.5.0
alembic==1.6.1
aniso8601==7.0.0
astroid==2.3.3
async-exit-stack==1.0.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
fastdiff==0.3.0
firebase-admin==5.0.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.0.0
grpcio==1.38.0
h11==0.12.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
jwcrypto==0.8
jws==0.1.3
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
mysql-connector-python==8.0.24
oauth2client==3.0.0
orjson==3.4.8
packaging==20.9
pluggy==0.13.1
promise==2.3
proto-plus==1.18.1
protobuf==3.15.8
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
PyJWT==2.1.0
pylint==2.4.4
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jwt==2.0.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.6.1
sip==4.19.8
six==1.15.0
snapshottest==0.6.0
SQLAlchemy==1.4.13
sseclient==0.0.27
starlette==0.13.6
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.26.3
uvicorn==0.13.3
wasmer==1.0.0
wasmer-compiler-cranelift==1.0.0
watchgod==0.6
websockets==8.1
wincertstore==0.2
wrapt==1.11.2
zipp==3.4.1
